# backend/internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: FuelOps Operator API
  version: "1.0"
  description: |
    Operator-facing REST API for the FuelOps Fuel Supply Intelligence & Resilience
    Platform. This service reads only from Postgres (the ingestor is the sole
    simulator client) and never calls the BUP Fuel Supply Simulator directly.
servers:
  - url: http://localhost:8000
    description: Local (docker compose)
paths:
  /api/state:
    get:
      summary: Latest world snapshot
      description: 'The most recent tick-fenced snapshot the ingestor persisted. Returns `{ "tick": null }` before the first snapshot exists.'
      operationId: getState
      responses:
        "200":
          description: Latest snapshot, or an empty marker when none exists yet.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/State"
                  - $ref: "#/components/schemas/EmptyState"
        "500":
          description: Database error.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /api/allocations/{id}/cancel:
    post:
      summary: Cancel a PENDING simulator allocation (the depot is refunded)
      operationId: cancelAllocation
      tags: [decisions]
      security: [{ bearer: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses: { "200": { $ref: "#/components/responses/Command" }, "404": { $ref: "#/components/responses/Error" }, "409": { $ref: "#/components/responses/Error" } }
  /api/simulate:
    post:
      summary: What-if for a candidate shipment
      description: Every constraint and simulator trap it would hit (route max, depot stock, per-depot dispatch cap, station room counting in-transit, route disrupted at departure, station outage) and its effect on stockout risk. Nothing is submitted.
      operationId: simulate
      tags: [decisions]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/Proposal" } } } }
      responses: { "200": { $ref: "#/components/responses/Json" } }
  /api/recommendations:
    get:
      summary: Recommendations for the current simulator epoch, newest first
      operationId: listRecommendations
      tags: [decisions]
      parameters:
        - { name: status, in: query, description: "Comma list: PROPOSED,APPROVED,REJECTED,SUBMITTED,EXPIRED,FAILED", schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer } }
      responses: { "200": { $ref: "#/components/responses/Json" } }
  /api/recommendations/{id}:
    get:
      summary: One recommendation with its evidence, decisions, submission state, live allocation and current validity
      operationId: getRecommendation
      tags: [decisions]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses: { "200": { $ref: "#/components/responses/Json" }, "404": { $ref: "#/components/responses/Error" } }
  /api/recommendations/{id}/explain:
    get:
      summary: Human-readable, operator-facing explanation of a recommendation (GenAI; deterministic hard numbers, LLM narrative with rule-based fallback)
      operationId: explainRecommendation
      tags: [decisions]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses: { "200": { $ref: "#/components/responses/Json" }, "404": { $ref: "#/components/responses/Error" } }
...
  /api/alerts/{id}/ack:
    post:
      summary: Acknowledge an alert
      operationId: ackAlert
      tags: [situation]
      security: [{ bearer: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses: { "200": { $ref: "#/components/responses/Json" }, "409": { $ref: "#/components/responses/Error" } }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: OPERATOR_TOKEN (approve, reject, allocate, cancel, ack) or ADMIN_TOKEN (all of that plus simulator control and policy). Reads need no token. With neither token configured, auth is off (local dev).
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: integer, minimum: 1 } }
  responses:
    Json: { description: OK, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
...
    Error: { description: Error, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
    Command: { description: "Command result (200 done, 202 still running — poll /api/admin/commands/{id}, 502 the simulator refused it).", content: { application/json: { schema: { $ref: "#/components/schemas/Command" } } } }
  schemas:
    ApiError:
      type: object
      properties:
        error:
          type: object
          properties: { code: { type: string, example: UNSAFE }, message: { type: string }, details: { type: array, items: { type: string } } }
          required: [code, message]
    Proposal:
      type: object
      properties:
        station_id: { type: string, example: station-mirpur }
        fuel_type: { type: string, enum: [DIESEL, PETROL, OCTANE] }
        route_id: { type: string, example: route-gazipur-mirpur }
        quantity: { type: number, example: 4000 }
      required: [station_id, fuel_type, route_id, quantity]
...
    Command:
      type: object
      properties:
        id: { type: integer }
        kind: { type: string, enum: [step, run, pause, reset, event, fault, faults_clear, cancel_allocation] }
        payload: { type: object, additionalProperties: true }
        status: { type: string, enum: [PENDING, DONE, FAILED] }
        result: { type: object, nullable: true, additionalProperties: true }
        error: { type: string, nullable: true }
        requested_by: { type: string }
        created_at: { type: string, format: date-time }
        done_at: { type: string, format: date-time, nullable: true }
    State:
      type: object
      properties:
        tick: { type: integer, example: 12 }
        stale: { type: boolean, description: "True if the snapshot was captured while a stale_data fault was active." }
        captured_at: { type: string, format: date-time }
        age_seconds: { type: number, format: double }
        snapshot: { type: object, description: "Raw sim.Snapshot payload.", additionalProperties: true }
      required: [tick, stale, captured_at, age_seconds, snapshot]
    EmptyState:
      type: object
      properties:
        tick: { type: "null" }
...
    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]